+++ CHANGELOG.md
@@ -8,9 +8,8 @@
 
 ## 24.3.1
 
-### ✨ Features and improvements
-
 ### 🐞 Bug fixes
+
 - The validator no longer fails if the style lacks a `glyphs` property ([#1347](https://github.com/maplibre/maplibre-style-spec/pull/1347))
 
 ## 24.3.0